refactor(site): move ItemList helper to its own module
buildTemplateItemListEntries serves the detail routes, not the index-page
card grid it lived beside; give it item-list.ts and repoint the callers.
