+++ site/src/pages/workflows/model/[name].astro
@@ -27,7 +27,7 @@ import { assertBrandSafe } from '../../../lib/workflow-pages/governance';
 import { buildModelPageMeta, isModelPageIndexable } from '../../../lib/workflow-pages/seo-page';
 import { assertNoOrphanedContent } from '../../../lib/workflow-pages/landing-content';
 import { buildFallbackPool } from '../../../lib/workflow-pages/fallback-pool';
-import { buildTemplateItemListEntries } from '../../../lib/workflow-pages/index-cards';
+import { buildTemplateItemListEntries } from '../../../lib/workflow-pages/item-list';
 import { resolveRailImages } from '../../../lib/workflow-pages/card-images';
 import {
   buildCollectionPageJsonLd,
