+++ site/src/pages/workflows/use-cases/[slug].astro
@@ -25,7 +25,7 @@ import {
 } from '../../../lib/hub-api';
 import { SEO_PAGES, type SeoPageDef } from '../../../lib/workflow-pages/use-cases';
 import { titleCaseKeyword } from '../../../lib/workflow-pages/schema';
-import { buildTemplateItemListEntries } from '../../../lib/workflow-pages/index-cards';
+import { buildTemplateItemListEntries } from '../../../lib/workflow-pages/item-list';
 import {
   resolveUseCasePageTemplates,
   relatedModelsForUseCase,
